# Conflicts:
#	consensus-module/poc/consensus-poc-base/src/main/java/io/nuls/consensus/poc/process/BlockProcess.java

+++ client-module/client/src/main/script/stop.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
-ps -ef|grep io.nuls.client.Bootstrap |grep -v grep |awk '{print $2}'|while read pid
-do
-   kill -9 $pid
-   echo "NULS has stopped."
-   break;
-done
+
+pid=`ps -ef| grep io.nuls.client.Bootstrap |grep -v 'grep' |awk '{print $2}'`
+
+if [ ! -z "$pid" ]; then
+    kill -9 $pid
+fi
